Separate AppDescription model from init
