+++ designsafe/apps/workspace/models/__init__.py
@@ -31,17 +31,3 @@ def _interactive_job_callback(sender, **kwargs):
         rp = RedisPublisher(facility = WEBSOCKETS_FACILITY, users=users)
         msg = RedisMessage(json.dumps(event_data))
         rp.publish_message(msg)
-
-
-class AppDescription(models.Model):
-    appId = models.CharField(max_length=120, unique=True, primary_key=True)
-    appDescription = models.TextField(help_text='App dropdown description text for apps that have a dropdown.')
-
-    def __str__(self):
-        return unicode(self).encode('utf-8')
-
-    def __unicode__(self):
-        return u"%s" % (self.appId)
-
-    def desc_to_dict(self):
-        return { 'appId': self.appId, 'appDescription': self.appDescription }
